# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mad Factory - Factory Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cashbox/summary:
    get:
      operationId: getCashboxSummary
      tags: [cashbox]
      summary: Get overall cashbox / financial summary
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cashbox summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashboxSummary"
  /cashbox/monthly-report:
    get:
      operationId: getMonthlyReport
      tags: [cashbox]
      summary: Monthly financial closing report
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly financial report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
  /cashbox/profit-loss:
    get:
      operationId: getProfitLossReport
      tags: [cashbox]
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Profit and loss report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitLossReport"
  /cashbox/closing:
    get:
      operationId: getCashboxClosing
      tags: [cashbox]
      summary: Daily/period cashbox close-out (worksheet-style)
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cashbox closing report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashboxClosingReport"
...
components:
  schemas:
...
    CashboxSummary:
      type: object
      properties:
        fromDate:
          type: ["string", "null"]
        toDate:
          type: ["string", "null"]
        totalSalesRevenue:
          type: number
        totalCustomerPayments:
          type: number
        totalPurchasesPaid:
          type: number
        totalSupplierPayments:
          type: number
        totalPayroll:
          type: number
        totalExpenses:
          type: number
        totalOtherIncome:
          type: number
        totalShareholderWithdrawals:
          type: number
        cashIn:
          type: number
        cashOut:
          type: number
        netCash:
          type: number
        totalCustomerDebt:
          type: number
        totalSupplierDebt:
          type: number
        latestExchangeRate:
          type: ["number", "null"]
      required:
        - totalSalesRevenue
        - totalCustomerPayments
        - totalPurchasesPaid
        - totalSupplierPayments
        - totalPayroll
        - totalExpenses
        - totalOtherIncome
        - totalShareholderWithdrawals
        - cashIn
        - cashOut
        - netCash
        - totalCustomerDebt
        - totalSupplierDebt

    # ── Monthly Report ────────────────────────────────────────────────────────
    MonthlyReport:
      type: object
      properties:
        month:
          type: string
        year:
          type: integer
        totalSales:
          type: number
        totalPurchases:
          type: number
        totalExpenses:
          type: number
        totalPayroll:
          type: number
        totalOtherIncome:
          type: number
        totalShareholderWithdrawals:
          type: number
        grossProfit:
          type: number
        netProfit:
          type: number
        cashIn:
          type: number
        cashOut:
          type: number
        netCash:
          type: number
      required:
        - month
        - year
        - totalSales
        - totalPurchases
        - totalExpenses
        - totalPayroll
        - totalOtherIncome
        - totalShareholderWithdrawals
        - grossProfit
        - netProfit
        - cashIn
        - cashOut
        - netCash

    # ── Profit/Loss Report ────────────────────────────────────────────────────
    ProfitLossReport:
      type: object
      properties:
        fromDate:
          type: ["string", "null"]
        toDate:
          type: ["string", "null"]
        totalRevenue:
          type: number
        totalCost:
          type: number
        totalPurchases:
          type: number
        cogsMissingCount:
          type: number
        cogsMissingRevenue:
          type: number
        grossProfit:
          type: number
        totalExpenses:
          type: number
        totalPayroll:
          type: number
        otherIncome:
          type: number
        shareholderWithdrawals:
          type: number
        netProfit:
          type: number
        profitByShareholdersBreakdown:
          type: array
          items:
            type: object
            properties:
              shareholderName:
                type: string
              sharePercentage:
                type: number
              profitShare:
                type: number
            required:
              - shareholderName
              - sharePercentage
              - profitShare
      required:
        - totalRevenue
        - totalCost
        - grossProfit
        - totalExpenses
        - totalPayroll
        - otherIncome
        - shareholderWithdrawals
        - netProfit
        - profitByShareholdersBreakdown

    # ── Cashbox Closing (worksheet-style) ─────────────────────────────────────
    CashboxClosingReport:
      type: object
      properties:
        fromDate: { type: ["string", "null"] }
        toDate: { type: ["string", "null"] }
        sales:
          type: object
          properties:
            gross: { type: number }
            discount: { type: number }
            net: { type: number }
            cashReceivedTotal: { type: number }
            cashReceivedCashVoucher: { type: number }
            cashReceivedInternalVoucher: { type: number }
            debtRemaining: { type: number }
        purchasesIqd:
          type: object
          properties:
            gross: { type: number }
            discount: { type: number }
            net: { type: number }
            cashPaidTotal: { type: number }
            cashPaidCashVoucher: { type: number }
            cashPaidInternalVoucher: { type: number }
            debtRemaining: { type: number }
        purchasesUsd:
          type: object
          properties:
            rate: { type: number }
            grossUsd: { type: number }
            discountUsd: { type: number }
            netUsd: { type: number }
            netIqd: { type: number }
            cashPaidUsd: { type: number }
            cashPaidIqd: { type: number }
            cashPaidCashVoucherUsd: { type: number }
            cashPaidInternalVoucherUsd: { type: number }
            debtRemainingUsd: { type: number }
            debtRemainingIqd: { type: number }
        payroll:
          type: object
          properties:
            due: { type: number }
            paid: { type: number }
            debt: { type: number }
        otherExpenses: { type: number }
        combined:
          type: object
          properties:
            totalExpenses: { type: number }
            totalCashPaid: { type: number }
            totalDebtPayable: { type: number }
        result:
          type: object
          properties:
            profit: { type: number }
            cashFlow: { type: number }
            netDebtChange: { type: number }
        shareholders:
          type: object
          properties:
            totalShares: { type: number }
            perShareValue: { type: number }
            breakdown:
              type: array
              items:
                type: object
                properties:
                  id: { type: number }
                  name: { type: string }
                  shareCount: { type: number }
                  sharePercentage: { type: number }
                  profitShare: { type: number }
        latestExchangeRate: { type: ["number", "null"] }

    # ── Treasury / Closings ───────────────────────────────────────────────────
...